% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gskmeans.R
\name{gskmeans}
\alias{gskmeans}
\title{Clustering with Geodesic Spherical k-Means Algorithm}
\usage{
gskmeans(x, k = 2, maxiter = 100)
}
\description{
Clustering with Geodesic Spherical k-Means Algorithm
}
