% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_gskmedoids.R
\name{gskmedoids}
\alias{gskmedoids}
\title{Clustering with Geodesic Spherical k-Medoids Algorithm}
\usage{
gskmedoids(x, k = 2, maxiter = 100)
}
\description{
Clustering with Geodesic Spherical k-Medoids Algorithm
}
