% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_vmf.R
\name{mle.vmf}
\alias{mle.vmf}
\title{MLE for von Mises-Fisher}
\usage{
mle.vmf(x, method = c("Banerjee", "Christie", "nmarg1", "nmarg2",
  "nmle1", "nmle2", "Song", "Sra", "Tanabe", "Uniroot"))
}
\description{
'Marginal Approximation (MA)' ; 1,2,3
'Marginal Exact         (ME)' ; 1,2,3
}
