% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_pdist.R
\name{pdist}
\alias{pdist}
\title{Pairwise Distance}
\usage{
pdist(x, mode = c("intrinsic", "extrinsic"), as.dist = FALSE)
}
\arguments{
\item{x}{(n x p) data matrix on S^{p-1}}
}
\description{
Pairwise Distance
}
