% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-RiemSphere.R
\docType{package}
\name{RiemSphere}
\alias{RiemSphere}
\alias{RiemSphere-package}
\title{Tools for Learning over Spheres}
\description{
Not Written Yet. what should I say. why not. testing.
some notable packages are
}
\details{
movMF
    rotasym
}
