% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.dpvmf.R
\name{sp.dpvmf}
\alias{sp.dpvmf}
\title{Dirichet Process Mixture of von Mises-Fisher Distributions}
\usage{
sp.dpvmf(data, kappa0)
}
\arguments{
\item{data}{an \eqn{(n\times d)} data matrix where \eqn{x_i \in \mathcal{S}^{d-1}}.}

\item{kappa0}{d}

\item{mu0}{d}

\item{a}{d}

\item{b}{d}

\item{w0}{d}

\item{mh.kappa}{stepsize for MH algorithm.}

\item{maxiter}{d}

\item{print.progress}{d}
}
\value{
a length-\code{maxiter} list whose elements are lists containing \describe{
\item{a}{b}
}
}
\description{
Dirichet Process Mixture of von Mises-Fisher Distributions
}
