% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.kmedoids.R
\name{sp.kmedoids}
\alias{sp.kmedoids}
\title{k-Medoids Clustering for Data on Hypersphere}
\usage{
sp.kmedoids(x, k = 2, type = c("intrinsic", "extrinsic"))
}
\description{
k-Medoids Clustering for Data on Hypersphere
}
