% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.mds.R
\name{sp.mds}
\alias{sp.mds}
\title{Multidimensional Scaling on Sphere}
\usage{
sp.mds(x, ndim = 2, type = c("intrinsic", "extrinsic"))
}
\description{
Multidimensional Scaling on Sphere
}
\examples{
## generate two-cluster data
mymu1 = c(0,0,0,1)  # center of class 1
mymu2 = c(-1,0,0,0) # center of class 2

x1 = rvmf(50, mymu1, kappa=15)
x2 = rvmf(50, mymu2, kappa=15)
xx = rbind(x1,x2)

## compute 2d embedding and visualization
mds2d <- sp.mds(xx, ndim=2)
plot(mds2d$embed[,1], mds2d$embed[,2], pch=19)

}
\references{
\insertRef{torgerson_multidimensional_1952}{DAS}
}
