% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_mixture_vmf.R
\name{sp.mixclust}
\alias{sp.mixclust}
\title{Clustering with Mixture of Distributions on Hypersphere}
\usage{
sp.mixclust(x, k = 2, n.start = 20, kernel = c("vmf", "spnorm"))
}
\description{
Clustering with Mixture of Distributions on Hypersphere
}
