% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.mixnorm.R
\name{sp.mixnorm}
\alias{sp.mixnorm}
\title{Mixture of Spherical Normal Distributions}
\usage{
sp.mixnorm(x, k = 2, n.start = 5, maxiter = 496, same.lambda = TRUE)
}
\description{
Mixture of Spherical Normal Distributions
}
\examples{
## generate two-cluster data
mymu1 = c(0,0,0,1)  # center of class 1
mymu2 = c(-1,0,0,0) # center of class 2
mymu3 = c(0,1,0,0)  # center of class 3

x1 = rvmf(50, mymu1, kappa=10)
x2 = rvmf(50, mymu2, kappa=10)
x3 = rvmf(50, mymu3, kappa=10)
xx = rbind(x1,x2,x3)

## apply clustering with different k values
sp.mixnorm(xx, k=2)

}
