% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.pdist.R
\name{sp.pdist}
\alias{sp.pdist}
\title{Pairwise Distance}
\usage{
sp.pdist(x, mode = c("intrinsic", "extrinsic"), as.dist = FALSE)
}
\arguments{
\item{x}{(n x p) data matrix on S^{p-1}}
}
\description{
Pairwise Distance
}
