% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.pdist2.R
\name{sp.pdist2}
\alias{sp.pdist2}
\title{Pairwise Distance of Two Sets of Data on Hypersphere}
\usage{
sp.pdist2(x1, x2, type = c("intrinsic", "extrinsic"))
}
\arguments{
\item{x1}{an \eqn{(m\times p)} row-stacked matrix for \eqn{\mathbb{S}^{p-1}}.}

\item{x2}{an \eqn{(n\times p)} row-stacked matrix for \eqn{\mathbb{S}^{p-1}}.}

\item{type}{type of distance, either \code{"intrinsic"} or \code{"extrinsic"}.}
}
\value{
an \eqn{(m \times n)} matrix of cross distances between \code{x1} and \code{x2}.
}
\description{
Pairwise Distance of Two Sets of Data on Hypersphere
}
