% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmean1_lrt.R
\name{spmean1.lrt}
\alias{spmean1.lrt}
\title{One-sample Location Test with Log-likelihood Ratio Test}
\usage{
spmean1.lrt(x, mu0, kappa0 = NULL)
}
\description{
One-sample Location Test with Log-likelihood Ratio Test
}
