% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_vmf.R
\name{vmf}
\alias{vmf}
\alias{dvmf}
\alias{rvmf}
\title{The von Mises-Fisher Distribution}
\usage{
dvmf(x, mu, kappa = 1, log = FALSE)

rvmf(n, mu, kappa = 1)
}
\arguments{
\item{x}{a param for toBar and notToBar}

\item{n}{the number

Could put here, but easiest if all params are
described in the same place, with page documentation, so none get repeated.}

\item{y}{a param just for notToBar}
}
\value{
Hard to have one return section for all functions,
might want to have a manual list here.
}
\description{
Group of functions Description section
}
\details{
Group of functions Details paragraph.
}
\section{After Arguments and Value sections}{

Despite its location, this actually comes after the Arguments and Value sections.
Also, don't need to use null, could annotate first function, and then
using function name as the groupBy name is more intuitive.
}

