% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp.mixnorm.R
\name{sp.mixnorm}
\alias{sp.mixnorm}
\title{Mixture of Spherical Normal Distributions}
\usage{
sp.mixnorm(x, k = 2, n.start = 5, maxiter = 496, same.lambda = TRUE)
}
\description{
Mixture of Spherical Normal Distributions
}
\examples{
## generate two-cluster data
mymu1 = c(0,0,0,1)  # center of class 1
mymu2 = c(-1,0,0,0) # center of class 2
mymu3 = c(0,1,0,0)  # center of class 3

x1 = rvmf(50, mymu1, kappa=10)
x2 = rvmf(50, mymu2, kappa=10)
x3 = rvmf(50, mymu3, kappa=10)
xx = rbind(x1,x2,x3)

## apply clustering with different k values
mix2 <- sp.kmeans(xx, k=2)
mix3 <- sp.kmeans(xx, k=3)
mix4 <- sp.kmeans(xx, k=4)

## compute 2-dimensional embedding for visualization
mds2d <- sp.mds(xx, ndim=2)
mdsx  <- mds2d$embed[,1]
mdsy  <- mds2d$embed[,2]

## compare via visualization
opar  <- par(mfrow=c(1,3), pty="s")
plot(mdsx, mdsy, col=mix2$cluster, main="k=2 means", pch=19)
plot(mdsx, mdsy, col=mix3$cluster, main="k=3 means", pch=19)
plot(mdsx, mdsy, col=mix4$cluster, main="k=4 means", pch=19)
par(opar)

}
